Fix SeasonalWindow naming conflict and MarketBasket comparingLong
- Rename 6-field SeasonalWindow (used by fetchProductSeasonalPerformance)
  to ProductSeasonalWindowDef to avoid clash with the 10-field SeasonalWindow
  used by fetchSeasonalCollections
- Fix MarketBasketService comparingInt → comparingLong for pairCount (long field)

+++ backend/src/main/java/com/pdv2cloud/service/MarketBasketService.java
@@ -186,7 +186,7 @@ public class MarketBasketService {
 
         // Ordena pela regra mais forte: pair_count desc, depois lift desc
         rules.sort(Comparator
-            .comparingInt(MarketBasketDTO::getPairCount).reversed()
+            .comparingLong(MarketBasketDTO::getPairCount).reversed()
             .thenComparingDouble(MarketBasketDTO::getLift).reversed()
             .thenComparingDouble(MarketBasketDTO::getConfidence).reversed()
         );
